GUI: Add functionality to "Crypt-it!" button.
Add an actionlistener to register "Crypt-it!"-button clicks.

When the "Crypt-it!" button is clicked and the textfields are not empty,
due to the added functionality user input will be passed as a parameter
to Crypter classes toMorse-method which will return crypted text as a string.

The returned string will be then set as a text to the right side textfield.

Signed-off-by: Roope Rantanen <roopepe@hotmail.com>