GUI: Add functionality and components to GUI and arrange them in a proper way.
Adds two new buttons to UI. The switch button moves the text from
the right side to the left side for further converting.
In addition the switch-button changes crypt-button to
decrypt-button and vice versa.The crypt-button calls Crypter objects
toMorse-method and decrypt-button calls toAlphabets-method from
the same object. Text will be then converted either to a code or
to original text.

Signed-off-by: Roope Rantanen <roopepe@hotmail.com>